requirements.txt
biopython>=1.79
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
xlwt>=1.3.0
PyYAML>=6.0
pytest>=7.0.0
pre-commit>=3.0.0 